images replace

+++ database/migrations/2021_02_11_180324_add_column_into_termostats.php
@@ -17,7 +17,7 @@ class AddColumnIntoTermostats extends Migration
             Schema::table('termostats', function (Blueprint $table) {
 
                 $table->integer('subdev_id')->nullable();
-                $table->double('current')->nullable()->change();
+                $table->float('current')->nullable()->change();
             });
         }
     }
